refactor: 결과 테스트 메서드 이동 ChallengerTest->ResultTest
